refactor: separate types into parse/type.ts file
